move safe environment impl into own file
